Change the Definition of response given by dropping the "mark-as-touched"-attempt completely.
Background:

The state of being done for an item is not well-defined in Verona2.
It pretty much open, what for example makes a checkbox-item `done`, when it's unchecked, since it can not be told if it's left unchecked intentional or not.

In the Verona2 version of this player, an item (wich is, spoken in HTML-terms any form element and any element with contenteditable-attribute), could be regarded as answered, but not valid. This lead to the question, how to tell if an unchecked checkbox oder an empty textbox was answered at all. The simple-player employed a mechanic, that marked an item as touched when it was entered by the mouse or git the focus. Touched items where regarded as answered, regardless if they had content or not, and regardless of their validity from the HTML point of view.

In Verona3, answers should not be propagated if they are invalid while "invalid" is defined as "not interesting for the study".

@see: https://github.com/verona-interfaces/player/issues/25

Form this followed that there is not 'complete-and-valid' state in the ResponseProgress anymore since it's the same as 'complete' and the Simple Player must not distinguish between answered and answered and also valid items anymore.
Additionally, the Simple Player should take raw HTML as literal as possible and the former "markAsTouched"-attempt could therefore lead to a behaviour that was not intended by the unit's author. So the mark-as-touched-attempt becomes obsolete and was removed.

The downside of this is, that the problem of intentionally unchecked checkboxes, wich is not solved in HTML (https://www.w3.org/html/wg/tracker/issues/111) in a satisfying way is not solved yet by the player in any way and remains for the unit's author.

Another issue left open is, how to handle element's pseudo-form-elements with `contenteditable`, that can not be marked as required (https://github.com/iqb-berlin/verona-player-simple/issues/8).

closes https://github.com/iqb-berlin/verona-player-simple/issues/6